Remove dependence on global cherrypy.tree
This is intended to achieve greater WSGI portability.
